requirements.txt
twilio~=9.2.3
Flask~=3.0.3
PyPDF2
python-dotenv~=1.0.1
langchain~=0.2.6
langchain_community
langchain_google_genai
requests
langchain-groq
google-cloud-aiplatform
groq
faiss-cpu
pypdf
langchain-core
werkzeug
python-dotenv~=1.0.1
streamlit
sumy